requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
numpy>=1.24.0
openai>=1.30.0
